Update links
There were a series of csv files in _data that controlled the links displayed in the header and footer. Some further info was controlled through site variables.

This commit draws all this into one place (the _config.yml) and uses YAML instead of CSV files to drive the output as it's easier to read for the uninitiated.