2021 rebuild (#2)
* Add files and folder to .gitignore
There are a number of files and folders used in the main site that shouldn’t be part of the theme underlying it, and a number of generic ignores copied across from the existing list.

* Add folders to .gitignore
Further folders from the main site that shouldn’t be part of the theme

* Update .gitignore

* Remove old files and folders
Much of this doesn’t need to be part of the theme folder

* Update .gitignore

* Add _config.yml
This will act as a useful template, so including in theme repo. It will be overridden when built by the _config in the main repo

* Add email to _config.yml

* Add explanatory text to config file

* Ignore the .lock file

* Update Gemfile
Include Jekyll plugins necessary for newer features

* Home page text removed from _config.yml
This text has never changed since the new site was created, so there’s no need to have this as a variable. Now hard coded into the site.

* Remove manuals from home page
This is for internal use so is very unlikely to be needed here.

* Remove includes for theme setup
This is out of date now, and was never used anyway.

* Manuals block no longer on home page
This was removed in a separate commit so it is no longer needed

* Update links

There were a series of csv files in _data that controlled the links displayed in the header and footer. Some further info was controlled through site variables.

This commit draws all this into one place (the _config.yml) and uses YAML instead of CSV files to drive the output as it's easier to read for the uninitiated.

* Remove padding
Excessive padding in the disclaimer area of the footer

* Adjust padding
Like goldilocks!

* Clean up the header

This commit:
- moves chunks of the <head> for the social graphs to separate files
- moves the 'rel=me' content to a separate file
- moves favicons to a separate file
- cleans up the conditional build logic so it's easier to read

* Update _config.yml

Minor formatting changes

* Amend structure of sass files

- Removed animations (empty file)
- Separated variables out into a default file and an override file

* Create pages collection

By addition a collection for pages, instead of creating stand files, we can avoid confusing _collections and normal folders, so that only essential files appear at the top of the hierarchy.

* Update to work with new _pages structure
There was a conflict declaring ‘collections’ in the front matter. Changing this to ‘loop’ fixes the issue.

* Change site.posts to site.news
We call this ‘news’ on the website, so changing the collection to match

* Fix social icons on home page

* Update topic pages social links
Social links now are read from the _config.yaml, reducing the amount of front-matter necessary in the topic pages.

(I can’t see a scenario where we would override these!)

* Update .gitignore

* Fix loop order
Posts were displaying in the wrong order

* Update README

* Update config file